backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==6.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Werkzeug==3.0.6
psycopg2-binary==2.9.7
python-dotenv==1.0.0
cryptography==43.0.3
pytest==7.4.2
pytest-cov==4.1.0
flake8==6.0.0

Flask-Limiter==3.5.0         # Rate limiting
python-magic==0.4.27         # File type validation (magic byte checking)
Pillow==10.3.0               # Image validation and processing
bleach>=6.0.0                # HTML sanitization for XSS protection
requests>=2.31.0             # CLI HTTP client for admin bulk upload

# TODO(security, JC): Adding these security packages eventually:
# PyJWT==2.8.0                 # JWT token authentication

APScheduler==3.11.0          # Scheduled deletions after 30 days
